refactoring and adding logging
