+++ BlackJack.Framework/Logging/Logger.cs
@@ -1,4 +1,5 @@
 ﻿using log4net;
+using log4net.Config;
 using System;
 
 namespace Game.Framework.Logging
@@ -7,6 +8,7 @@ namespace Game.Framework.Logging
     {
         public Logger()
         {
+            XmlConfigurator.Configure();
             _log = LogManager.GetLogger("Game.Logger");
         }
 
@@ -50,7 +52,7 @@ namespace Game.Framework.Logging
 
         public void Debug(string data)
         {
-            if (data != null && IsDebugEnabled)
+            if (IsDebugEnabled)
             {
                 _log.Debug(data);
             }
@@ -58,26 +60,33 @@ namespace Game.Framework.Logging
 
         public void Info(string data)
         {
-            if (data != null && IsInfoEnabled)
+            if (IsInfoEnabled)
             {
                 _log.Info(data);
             }
         }
 
-        public void Error(string data, Exception exception)
+        public void Error(string data, Exception exception = null)
         {
-            if (data != null && IsErrorEnabled)
+            if (IsErrorEnabled)
             {
                 if (exception != null)
                 {
-                    _log.Error($" [Exception]: {exception.ToString()} [Message]: {data}");
+                    _log.Error($" [Exception]: {exception} [Exception Message]: {exception.Message} [Message]: {data}");
+                }
+                else
+                {
+                    _log.Error(data);
                 }
             }
         }
 
         public void Warning(string data)
         {
-            throw new NotImplementedException();
+            if (IsWarnEnabled)
+            {
+                _log.Warn(data);
+            }
         }
     }
 }
