+++ BlackJack.persistence/Repositories/GameUserRepository.cs
@@ -1,6 +1,7 @@
-﻿using System;
+﻿using Game.Models.Entities;
+using System;
 using System.Data.Entity.Infrastructure;
-using static Game.Core.Enumerations.EnumBag;
+using static Game.Models.Enumerations.EnumBag;
 
 namespace Game.Persistence.Repositories
 {
