+++ BlackJack.persistence/Repository.cs
@@ -1,11 +1,11 @@
-﻿using Game.Core.Entities;
+﻿using Game.Models.Entities;
 using System;
 using System.Collections.Generic;
 using System.Data.Entity;
 using System.Data.Entity.Infrastructure;
 using System.Linq;
 using System.Linq.Expressions;
-using static Game.Core.Enumerations.EnumBag;
+using static Game.Models.Enumerations.EnumBag;
 
 namespace Game.Persistence
 {
