+++ Blackjack.client/app.config
@@ -1,6 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <configuration>
   <configSections>
+    <section name="log4net" type="log4net.Config.Log4NetConfigurationSectionHandler, log4net"/>
     <!-- For more information on Entity Framework configuration, visit http://go.microsoft.com/fwlink/?LinkID=237468 -->
     <section name="entityFramework" type="System.Data.Entity.Internal.ConfigFile.EntityFrameworkSection, EntityFramework, Version=6.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" requirePermission="false" />
   </configSections>
@@ -14,5 +15,30 @@
   <connectionStrings>
     <add name="gamedb" providerName="System.Data.SqlClient" connectionString="Server=tcp:crouse01.database.windows.net,1433;Initial Catalog=gamedb;Persist Security Info=False;User ID=aldos;Password=4ld0sSql/;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=30;"/>
   </connectionStrings>
+
+  <log4net>
+    
+    <logger name="Game.Logger">
+      <level value="ALL" />
+      <appender-ref ref="ConsoleAppender" />
+      <appender-ref ref="FileAppender" />
+    </logger>
+    
+    <!-- appenders go here -->
+    <appender name="ConsoleAppender" type="log4net.Appender.ConsoleAppender">
+      <layout type="log4net.Layout.PatternLayout">
+        <conversionPattern value="%date %level %logger - %message%newline" />
+      </layout>
+    </appender>
+    <appender name="FileAppender" type="log4net.Appender.FileAppender">
+      <file type="log4net.Util.PatternString" value="C:\GameLogs\TraceLog.%property{log4net:HostName}.txt" />
+      <appendToFile value="true" />
+      <lockingModel type="log4net.Appender.FileAppender+MinimalLock" />
+      <layout type="log4net.Layout.PatternLayout">
+        <conversionPattern value="%date %level %logger - %message%newline" />
+      </layout>
+    </appender> 
+    
+  </log4net>
 
 </configuration>
