+++ Blackjack.models/Services/ServiceManager.cs
@@ -1,6 +1,7 @@
 ﻿using Game.Core.Requests;
 using Game.Core.Response;
 using Game.Framework.Extensions;
+using Game.Framework.Logging;
 using System;
 using System.Diagnostics;
 
@@ -8,6 +9,13 @@ namespace Game.Core.Services
 {
     public class ServiceManager
     {
+        public ServiceManager(ILogger logger)
+        {
+            Logger = logger;
+        }
+
+        protected ILogger Logger { get; set; }
+
         protected TReturn Execute<TReturn>(IValidateableRequest request, Action<TReturn> action)
             where TReturn : INotificationResponse, new()
         {
@@ -29,13 +37,15 @@ namespace Game.Core.Services
                 }
                 action.Invoke(response);
             }
-            catch(Exception e)
+            catch(Exception ex)
             {
-                response.Notifications.AddError($"An unexpected error occured: { e.ToString() }");
+                Logger.Error($"[Class]: { GetType() } [Method]: { (new StackTrace()).GetFrame(1).GetMethod().Name }. ", ex);
+                response.Notifications.AddError($"An unexpected error occured: { ex.Message }.");
             }
             finally
             {
-                stopwatch.Stop();               
+                stopwatch.Stop();
+                Logger.Info($"[Class]: { GetType() }. [Method]: { (new StackTrace()).GetFrame(1).GetMethod().Name }. [CompletionTime]: {stopwatch.ElapsedMilliseconds}ms ");                             
             }
             return response;
         }
