+++ Game.Models/Entities/IEntity.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using static Game.Models.Enumerations.EnumBag;
 
 namespace Game.Models.Entities
 {
